requirements.txt
pytest~=8.1.1
pytest-asyncio~=0.23.6
aiohttp~=3.9.3
pytest-mock~=3.14.0
asyncmock~=0.4.2
requests>=2.0.0
python-dotenv
flake8
datetime